requirements.txt
raylib-py==5.5.0
torch==2.2.2
jdata==0.6.0
openexr==3.3.4
mitsuba==3.6.4
mitransient==1.1.2